Move default legend in its own file
