+++ packages/x-charts/src/ChartsLegend/index.ts
@@ -1,4 +1,5 @@
 export * from './ChartsLegend';
+export * from './DefaultChartsLegend';
 export * from './chartsLegendClasses';
 
 export * from './utils';
